requirements_pip.txt
BeautifulSoup==3.2.0
Django==1.6.5
DjangoSEO==1.0
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
PyPDF2==1.25.1
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
app-version==0.2.1
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
django-appconf==1.0.1
django-compressor==1.5
django-endless-pagination==2.0
django-extensions==1.5.5
django-haystack==2.1.0
django-ipware==1.1.0
django-nose==1.4
django-oauth-plus==2.0
django-pagination==1.0.7
django-paypal==0.1.2
django-recaptcha==0.0.6
django-simple-captcha==0.3.7
django-smtp-ssl==1.0
django-social-auth==0.7.17
django-sorting==0.1
django-tagging==0.3.1
django-templated-email==0.4.7
django-thumbnailfield==0.2.5
djrill==0.9.0
dnspython==1.11.1
duplicity==0.6.23
elasticstack==0.1.1
geopy==0.98
gunicorn==19.3.0
html5lib==0.999
httplib2==0.9.1
ipgetter==0.6
lockfile==0.8
lxml==3.3.3
mysql-connector-python==1.1.6
mysql-utilities==1.3.5
nose==1.3.7
oauth2==1.5.211
oauthlib==0.6.1
oneconf==0.3.7
openpyxl==1.6.1
paramiko==1.10.1
pbr==1.4.0
pexpect==3.1
pisa==3.0.32
piston-mini-client==0.7.5
py-moneyed==0.6.0
pyOpenSSL==0.13
pyPdf==1.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pyelasticsearch==0.6.1
pygobject==3.12.0
pyserial==2.6
pysmbc==1.0.14.1
pysolr==3.1.0
pysqlite==2.6.3
python-apt==0.9.3.5
python-dateutil==1.5
python-debian==0.1.21-nmu2ubuntu2
python-openid==2.2.5
pytz==2015.4
pyxdg==0.25
recaptcha-client==1.0.6
reportlab==2.5
requests==2.7.0
sessioninstaller==0.0.0
simplejson==3.8.0
six==1.9.0
software-center-aptd-plugins==0.0.0
standardiser==0.1.7
stevedore==1.7.0
system-service==0.1.6
tolerance==0.1.2
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==13.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
xhtml2pdf==0.0.6
xlwt==0.7.5
zope.interface==4.0.5
